chore(conflicts): resolved merge conflicts with develop branch
[Finishes #167733644]
