+++ package.json
@@ -29,7 +29,6 @@
     "@sendgrid/mail": "^6.4.0",
     "bcryptjs": "^2.4.3",
     "@hapi/joi": "^15.1.1",
-    "body-parser": "^1.18.3",
     "cors": "^2.8.4",
     "cross-env": "^5.2.0",
     "dotenv": "^8.0.0",
